Added a first version of the BigSet.
The BigSet is inspired by the already existing OR-Set, only that it spreads its elements over multiple shards. Each element is hashed by using Erlang's phash2/2 with range 0-2^X. X can be set by the user upon creation of the Bigset by giving or is just set to 16 if no parameter is given. If more than Y (also set by the user or set to 30 if no parameter given) elements are added to a shard, it is split in two new shards, which both have exactly half the range of the old shard.

The shards are organized in a tree, as can be seen in antidote_crdt_bigset_shardtree, where the inner nodes are for navigation only (hash value of the element and key of the node are compared), and the leaves contain the current shards. When adding or removing an element, the tree is searched for the fitting shard. Removing enough elements from a shard can lead to that shard merging with its sibling, if that one is also a leaf. In that case, both shards are merged, the leaves are dropped and their parent is replaced with the merged shard.